moved pages dir into src dir
