+++ client/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client'
 import { BrowserRouter } from "react-router-dom";
-import App from './app';
+import App from './src/app';
 import AppContextProvider from './appState/index.js';
 
 const Stack =
